+++ e-invoice-java-client-okhttp/src/main/kotlin/com/e_invoice/api/client/okhttp/EInvoiceOkHttpClient.kt
@@ -31,8 +31,6 @@ class EInvoiceOkHttpClient private constructor() {
         private var timeout: Timeout = Timeout.default()
         private var proxy: Proxy? = null
 
-        fun development() = apply { baseUrl(ClientOptions.DEVELOPMENT_URL) }
-
         fun baseUrl(baseUrl: String) = apply { clientOptions.baseUrl(baseUrl) }
 
         /**
@@ -168,13 +166,7 @@ class EInvoiceOkHttpClient private constructor() {
         fun build(): EInvoiceClient =
             EInvoiceClientImpl(
                 clientOptions
-                    .httpClient(
-                        OkHttpClient.builder()
-                            .baseUrl(clientOptions.baseUrl())
-                            .timeout(timeout)
-                            .proxy(proxy)
-                            .build()
-                    )
+                    .httpClient(OkHttpClient.builder().timeout(timeout).proxy(proxy).build())
                     .build()
             )
     }