+++ e-invoice-java-core/src/main/kotlin/com/e_invoice/api/services/async/OutboxServiceAsyncImpl.kt
@@ -20,6 +20,7 @@ import com.e_invoice.api.models.outbox.OutboxListDraftDocumentsParams
 import com.e_invoice.api.models.outbox.OutboxListReceivedDocumentsPageAsync
 import com.e_invoice.api.models.outbox.OutboxListReceivedDocumentsParams
 import java.util.concurrent.CompletableFuture
+import java.util.function.Consumer
 
 class OutboxServiceAsyncImpl internal constructor(private val clientOptions: ClientOptions) :
     OutboxServiceAsync {
@@ -29,6 +30,9 @@ class OutboxServiceAsyncImpl internal constructor(private val clientOptions: Cli
     }
 
     override fun withRawResponse(): OutboxServiceAsync.WithRawResponse = withRawResponse
+
+    override fun withOptions(modifier: Consumer<ClientOptions.Builder>): OutboxServiceAsync =
+        OutboxServiceAsyncImpl(clientOptions.toBuilder().apply(modifier::accept).build())
 
     override fun listDraftDocuments(
         params: OutboxListDraftDocumentsParams,
@@ -49,6 +53,13 @@ class OutboxServiceAsyncImpl internal constructor(private val clientOptions: Cli
 
         private val errorHandler: Handler<JsonValue> = errorHandler(clientOptions.jsonMapper)
 
+        override fun withOptions(
+            modifier: Consumer<ClientOptions.Builder>
+        ): OutboxServiceAsync.WithRawResponse =
+            OutboxServiceAsyncImpl.WithRawResponseImpl(
+                clientOptions.toBuilder().apply(modifier::accept).build()
+            )
+
         private val listDraftDocumentsHandler: Handler<PaginatedDocumentResponse> =
             jsonHandler<PaginatedDocumentResponse>(clientOptions.jsonMapper)
                 .withErrorHandler(errorHandler)
@@ -60,6 +71,7 @@ class OutboxServiceAsyncImpl internal constructor(private val clientOptions: Cli
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.GET)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("api", "outbox", "drafts")
                     .build()
                     .prepareAsync(clientOptions, params)
@@ -98,6 +110,7 @@ class OutboxServiceAsyncImpl internal constructor(private val clientOptions: Cli
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.GET)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("api", "outbox", "")
                     .build()
                     .prepareAsync(clientOptions, params)