+++ e-invoice-java-core/src/main/kotlin/com/e_invoice/api/services/async/WebhookServiceAsync.kt
@@ -2,6 +2,7 @@
 
 package com.e_invoice.api.services.async
 
+import com.e_invoice.api.core.ClientOptions
 import com.e_invoice.api.core.RequestOptions
 import com.e_invoice.api.core.http.HttpResponseFor
 import com.e_invoice.api.models.webhooks.WebhookCreateParams
@@ -12,6 +13,7 @@ import com.e_invoice.api.models.webhooks.WebhookResponse
 import com.e_invoice.api.models.webhooks.WebhookRetrieveParams
 import com.e_invoice.api.models.webhooks.WebhookUpdateParams
 import java.util.concurrent.CompletableFuture
+import java.util.function.Consumer
 
 interface WebhookServiceAsync {
 
@@ -19,6 +21,13 @@ interface WebhookServiceAsync {
      * Returns a view of this service that provides access to raw HTTP responses for each method.
      */
     fun withRawResponse(): WithRawResponse
+
+    /**
+     * Returns a view of this service with the given option modifications applied.
+     *
+     * The original service is not modified.
+     */
+    fun withOptions(modifier: Consumer<ClientOptions.Builder>): WebhookServiceAsync
 
     /** Create a new webhook */
     fun create(params: WebhookCreateParams): CompletableFuture<WebhookResponse> =
@@ -157,6 +166,15 @@ interface WebhookServiceAsync {
      * A view of [WebhookServiceAsync] that provides access to raw HTTP responses for each method.
      */
     interface WithRawResponse {
+
+        /**
+         * Returns a view of this service with the given option modifications applied.
+         *
+         * The original service is not modified.
+         */
+        fun withOptions(
+            modifier: Consumer<ClientOptions.Builder>
+        ): WebhookServiceAsync.WithRawResponse
 
         /**
          * Returns a raw HTTP response for `post /api/webhooks/`, but is otherwise the same as