+++ e-invoice-java-core/src/main/kotlin/com/e_invoice/api/services/async/WebhookServiceAsyncImpl.kt
@@ -24,6 +24,7 @@ import com.e_invoice.api.models.webhooks.WebhookResponse
 import com.e_invoice.api.models.webhooks.WebhookRetrieveParams
 import com.e_invoice.api.models.webhooks.WebhookUpdateParams
 import java.util.concurrent.CompletableFuture
+import java.util.function.Consumer
 import kotlin.jvm.optionals.getOrNull
 
 class WebhookServiceAsyncImpl internal constructor(private val clientOptions: ClientOptions) :
@@ -34,6 +35,9 @@ class WebhookServiceAsyncImpl internal constructor(private val clientOptions: Cl
     }
 
     override fun withRawResponse(): WebhookServiceAsync.WithRawResponse = withRawResponse
+
+    override fun withOptions(modifier: Consumer<ClientOptions.Builder>): WebhookServiceAsync =
+        WebhookServiceAsyncImpl(clientOptions.toBuilder().apply(modifier::accept).build())
 
     override fun create(
         params: WebhookCreateParams,
@@ -75,6 +79,13 @@ class WebhookServiceAsyncImpl internal constructor(private val clientOptions: Cl
 
         private val errorHandler: Handler<JsonValue> = errorHandler(clientOptions.jsonMapper)
 
+        override fun withOptions(
+            modifier: Consumer<ClientOptions.Builder>
+        ): WebhookServiceAsync.WithRawResponse =
+            WebhookServiceAsyncImpl.WithRawResponseImpl(
+                clientOptions.toBuilder().apply(modifier::accept).build()
+            )
+
         private val createHandler: Handler<WebhookResponse> =
             jsonHandler<WebhookResponse>(clientOptions.jsonMapper).withErrorHandler(errorHandler)
 
@@ -85,6 +96,7 @@ class WebhookServiceAsyncImpl internal constructor(private val clientOptions: Cl
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.POST)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("api", "webhooks", "")
                     .body(json(clientOptions.jsonMapper, params._body()))
                     .build()
@@ -118,6 +130,7 @@ class WebhookServiceAsyncImpl internal constructor(private val clientOptions: Cl
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.GET)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("api", "webhooks", params._pathParam(0))
                     .build()
                     .prepareAsync(clientOptions, params)
@@ -150,6 +163,7 @@ class WebhookServiceAsyncImpl internal constructor(private val clientOptions: Cl
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.PUT)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("api", "webhooks", params._pathParam(0))
                     .body(json(clientOptions.jsonMapper, params._body()))
                     .build()
@@ -181,6 +195,7 @@ class WebhookServiceAsyncImpl internal constructor(private val clientOptions: Cl
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.GET)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("api", "webhooks", "")
                     .build()
                     .prepareAsync(clientOptions, params)
@@ -214,6 +229,7 @@ class WebhookServiceAsyncImpl internal constructor(private val clientOptions: Cl
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.DELETE)
+                    .baseUrl(clientOptions.baseUrl())
                     .addPathSegments("api", "webhooks", params._pathParam(0))
                     .apply { params._body().ifPresent { body(json(clientOptions.jsonMapper, it)) } }
                     .build()