+++ e-invoice-java-core/src/main/kotlin/com/e_invoice/api/services/blocking/documents/UblService.kt
@@ -2,11 +2,13 @@
 
 package com.e_invoice.api.services.blocking.documents
 
+import com.e_invoice.api.core.ClientOptions
 import com.e_invoice.api.core.RequestOptions
 import com.e_invoice.api.core.http.HttpResponseFor
 import com.e_invoice.api.models.documents.ubl.UblGetParams
 import com.e_invoice.api.models.documents.ubl.UblGetResponse
 import com.google.errorprone.annotations.MustBeClosed
+import java.util.function.Consumer
 
 interface UblService {
 
@@ -14,6 +16,13 @@ interface UblService {
      * Returns a view of this service that provides access to raw HTTP responses for each method.
      */
     fun withRawResponse(): WithRawResponse
+
+    /**
+     * Returns a view of this service with the given option modifications applied.
+     *
+     * The original service is not modified.
+     */
+    fun withOptions(modifier: Consumer<ClientOptions.Builder>): UblService
 
     /** Get the UBL for an invoice or credit note */
     fun get(documentId: String): UblGetResponse = get(documentId, UblGetParams.none())
@@ -44,6 +53,13 @@ interface UblService {
 
     /** A view of [UblService] that provides access to raw HTTP responses for each method. */
     interface WithRawResponse {
+
+        /**
+         * Returns a view of this service with the given option modifications applied.
+         *
+         * The original service is not modified.
+         */
+        fun withOptions(modifier: Consumer<ClientOptions.Builder>): UblService.WithRawResponse
 
         /**
          * Returns a raw HTTP response for `get /api/documents/{document_id}/ubl`, but is otherwise