+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.1.0-alpha.2 (2025-06-13)
+
+Full Changelog: [v0.1.0-alpha.1...v0.1.0-alpha.2](https://github.com/e-invoice-be/e-invoice-java/compare/v0.1.0-alpha.1...v0.1.0-alpha.2)
+
+### Features
+
+* **api:** api update ([315c958](https://github.com/e-invoice-be/e-invoice-java/commit/315c9588a2b527ade8df91c9d143ae16168e4c6a))
+* **api:** manual updates ([0c05809](https://github.com/e-invoice-be/e-invoice-java/commit/0c0580984c881990dc8093a9e3e64cdf89b2e05c))
+* **api:** manual updates ([97bf1a6](https://github.com/e-invoice-be/e-invoice-java/commit/97bf1a6778519ef3d24dc8b8c5a73dbf994c31ee))
+* **client:** add a `withOptions` method ([46b8e36](https://github.com/e-invoice-be/e-invoice-java/commit/46b8e36c4613c4cae250249ded9989667d6d4904))
+* **client:** implement per-endpoint base URL support ([6aad447](https://github.com/e-invoice-be/e-invoice-java/commit/6aad447f90c5079e3fd844e3724f02ed1b7acc17))
+
 ## 0.1.0-alpha.1 (2025-06-01)
 
 Full Changelog: [v0.0.1-alpha.0...v0.1.0-alpha.1](https://github.com/e-invoice-be/e-invoice-java/compare/v0.0.1-alpha.0...v0.1.0-alpha.1)